RCTConvert: remove deprecation matrix input warning.
A warning pops up every time I navigate to a new screen, only in remote
debug. I suspect it is the ExNavigator we use, because I saw it has
code related to "transformMatrix" which is the deprecated property.
Until we move to a newer navigator, I remove this deprecation warning.